allow building multicall binary as dynamic library
