+++ README.md
@@ -1,6 +1,6 @@
 <!-- markdownlint-disable MD033 MD041 MD002 -->
 <!-- markdownlint-disable commands-show-output no-duplicate-heading -->
-<!-- spell-checker:ignore markdownlint ; (options) DESTDIR UTILNAME manpages reimplementation oranda -->
+<!-- spell-checker:ignore markdownlint ; (options) DESTDIR UTILNAME dylib manpages reimplementation oranda -->
 <div class="oranda-hide">
 <div align="center">
 
@@ -129,6 +129,22 @@ the `--package` [aka `-p`] option). For example:
 cargo build -p uu_base32 -p uu_cat -p uu_echo -p uu_rm
 ```
 
+The multicall code can also be built as a dynamic library, and the multicall
+binary turned into a minimal shim to load it:
+
+```shell
+cargo rustc --release --features dynamic,unix --lib --crate-type dylib
+cargo build --release --features dynamic
+```
+
+Note that the shim multicall binary will not function unless the library is
+accessible in one of the dynamic loader's search paths—either a default library
+search path, or a modified environment. For example, on Linux:
+
+```shell
+LD_LIBRARY_PATH=$(pwd)/target/release/ ./target/release/coreutils
+```
+
 ### GNU Make
 
 Building using `make` is a simple process as well.
@@ -212,11 +228,24 @@ To install the multicall binary:
 make MULTICALL=y install
 ```
 
-Set install parent directory (default value is /usr/local):
+Set install parent directory (default value is `/usr/local`):
 
 ```shell
 # DESTDIR is also supported
 make PREFIX=/my/path install
+```
+
+The dynamically liked multicall library and binary can be installed instead.
+Note, however, that while the standard local `/usr/local/bin` path for binaries
+is typically in the default `PATH` environment variable, the standard local
+`/usr/local/lib` path for libraries is typically *not* in the default library
+search paths. If you do not address this in a persistent manner, then do a
+default dynamic install, your system will likely to fail to fully boot. With
+that in mind, to install the dynamically linked library and shim multicall
+binary:
+
+```shell
+make MULTICALL=y DYNAMIC=y install
 ```
 
 Installing with `make` installs shell completions for all installed utilities
@@ -293,6 +322,12 @@ To uninstall the multicall binary:
 
 ```shell
 make MULTICALL=y uninstall
+```
+
+Similarly, for the dynamically linked library and binary:
+
+```shell
+make MULTICALL=y DYNAMIC=y uninstall
 ```
 
 To uninstall from a custom parent directory:
